Datetime casting with custom format (#2658)
As mentioned in the #2655 issue, casting date types with a custom format did not work properly. So I decided to cover all date types and fix this issue.
I tested these date types:

- immutable_date
- immutable_date with custom formatting
- immutable_datetime
- immutable_datetime with custom formatting
To achieve this goal, I override one of the cases in castAttribute method to handle immutable_date date type. Also, I override transformModelValue method to check if the result is a DateTimeInterface or not for applying defined custom formatting on the Carbon instance and resetting time in a custom date type.
In the end, I should say that all date values will be stored in DB as Date type. (not as an object or string anymore)

---------

Co-authored-by: Jérôme Tamarelle <jerome@tamarelle.net>